+++ tiger-proxy/src/test/java/de/gematik/test/tiger/proxy/controller/TigerWebUiControllerTest.java
@@ -16,9 +16,11 @@ import de.gematik.rbellogger.data.facet.RbelMessageTimingFacet;
 import de.gematik.test.tiger.proxy.TigerProxy;
 import io.restassured.RestAssured;
 import java.time.OffsetDateTime;
+import kong.unirest.JsonNode;
 import kong.unirest.Unirest;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
+import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
@@ -28,7 +30,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
 
 @Slf4j
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
+    properties = "tigerProxy.skipDisplayWhenMessageLargerThanKb = 1")
 public class TigerWebUiControllerTest {
 
     public static WireMockServer fakeBackendServer;
@@ -78,7 +81,7 @@ public class TigerWebUiControllerTest {
     }
 
     @Test
-    public void checkHtmlIsReturned() {
+    void checkHtmlIsReturned() {
         RestAssured.given().get(getWebUiUrl())
             .then()
             .statusCode(200)
@@ -86,7 +89,7 @@ public class TigerWebUiControllerTest {
     }
 
     @Test
-    public void checkMsgIsReturned() {
+    void checkMsgIsReturned() {
         RestAssured.given().get(getWebUiUrl() + "/getMsgAfter")
             .then()
             .statusCode(200)
@@ -96,7 +99,7 @@ public class TigerWebUiControllerTest {
     }
 
     @Test
-    public void checkOnlyOneMsgIsReturnedWithLastMsgUuidSupplied() {
+    void checkOnlyOneMsgIsReturnedWithLastMsgUuidSupplied() {
         RestAssured.given()
             .get(getWebUiUrl() + "/getMsgAfter?lastMsgUuid=" + tigerProxy.getRbelMessagesList().get(0).getUuid())
             .then()
@@ -106,7 +109,7 @@ public class TigerWebUiControllerTest {
     }
 
     @Test
-    public void checkAllTrafficSuppliedWhenDownloadWithoutFilteredUuids() {
+    void checkAllTrafficSuppliedWhenDownloadWithoutFilteredUuids() {
         RestAssured.given().get(getWebUiUrl() + "/trafficLog.tgr")
             .then()
             .statusCode(200)
@@ -115,7 +118,7 @@ public class TigerWebUiControllerTest {
     }
 
     @Test
-    public void checkSuppliedUuidsAreFilteredOutWhenDownloadingTraffic() {
+    void checkSuppliedUuidsAreFilteredOutWhenDownloadingTraffic() {
         RestAssured.given()
             .get(getWebUiUrl() + "/trafficLog.tgr?lastMsgUuid=" + tigerProxy.getRbelMessagesList().get(0).getUuid())
             .then()
@@ -124,7 +127,7 @@ public class TigerWebUiControllerTest {
     }
 
     @Test
-    public void checkNoMsgIsReturnedIfNoneExistsAfterRequested() {
+    void checkNoMsgIsReturnedIfNoneExistsAfterRequested() {
         RestAssured.given()
             .get(getWebUiUrl() + "/getMsgAfter?lastMsgUuid=" + tigerProxy.getRbelMessagesList().get(1).getUuid())
             .then()
@@ -134,7 +137,7 @@ public class TigerWebUiControllerTest {
 
 
     @Test
-    public void checkNoMsgIsReturnedAfterReset() {
+    void checkNoMsgIsReturnedAfterReset() {
         RestAssured.given().get(getWebUiUrl() + "/resetMsgs")
             .then()
             .statusCode(200);
@@ -146,7 +149,7 @@ public class TigerWebUiControllerTest {
     }
 
     @Test
-    public void checkCorrectMenuStringsAreSupplied() {
+    void checkCorrectMenuStringsAreSupplied() {
         RestAssured.given()
             .get(getWebUiUrl() + "/getMsgAfter")
             .then()
@@ -190,7 +193,7 @@ public class TigerWebUiControllerTest {
     }
 
     @Test
-    public void filterOutResponses_shouldStillAppearInPairs() {
+    void filterOutResponses_shouldStillAppearInPairs() {
         RestAssured.given()
             .get(getWebUiUrl() + "/getMsgAfter?filterCriterion=isRequest")
             .then()
@@ -200,13 +203,28 @@ public class TigerWebUiControllerTest {
 
     @Test
     void filterOutRequests_shouldStillAppearInPairs() {
-        tigerProxy.getRbelMessages()
-            .forEach(e -> System.out.println(e.printTreeStructure()));
-
         RestAssured.given()
             .get(getWebUiUrl() + "/getMsgAfter?filterCriterion=isResponse")
             .then()
             .statusCode(200)
             .body("metaMsgList.size()", equalTo(2));
+    }
+
+    @Test
+    void largeMessage_shouldNotBeRenderedCompletelyButStillAppear() {
+        tigerProxy.getRbelMessages().clear();
+
+        val proxyRest = Unirest.spawnInstance();
+        proxyRest.config().proxy("localhost", tigerProxy.getProxyPort());
+        final String longString = RandomStringUtils.randomAlphanumeric(2_000);
+        proxyRest.post("http://localhost:" + fakeBackendServer.port() + "/foobar")
+            .body("{'randomStringForLulz':'" + longString + "'}")
+            .asString();
+
+        final JsonNode body = Unirest.get(getWebUiUrl() + "/getMsgAfter").asJson().getBody();
+        System.out.println(body.toString());
+        assertThat(body.getObject().getJSONArray("htmlMsgList").getJSONObject(0).getString("html"))
+            .contains("foobar")
+            .doesNotContain(longString);
     }
 }