+++ tiger-rbel/src/main/java/de/gematik/rbellogger/RbelLogger.java
@@ -48,7 +48,7 @@ public class RbelLogger {
             .rbelKeyManager(new RbelKeyManager())
             .manageBuffer(configuration.isManageBuffer())
             .rbelBufferSizeInMb(configuration.getRbelBufferSizeInMb())
-            .skipParsingWhenMessageLargerThanMb(configuration.getSkipParsingWhenMessageLargerThanMb())
+            .skipParsingWhenMessageLargerThanKb(configuration.getSkipParsingWhenMessageLargerThanKb())
             .build();
 
         rbelConverter.registerListener(new RbelX5cKeyReader());