+++ tiger-rbel/src/main/java/de/gematik/rbellogger/configuration/RbelConfiguration.java
@@ -43,7 +43,7 @@ public class RbelConfiguration {
     @Builder.Default
     private int rbelBufferSizeInMb = 1024;
     @Builder.Default
-    private int skipParsingWhenMessageLargerThanMb = 16;
+    private int skipParsingWhenMessageLargerThanKb = 16_000;
     @Builder.Default
     private boolean manageBuffer = false;