+++ tiger-rbel/src/test/java/de/gematik/rbellogger/converter/RbelMtomConverterTest.java
@@ -22,7 +22,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class RbelMtomConverterTest {
+class RbelMtomConverterTest {
 
     private RbelLogger rbelLogger;